# Improve performance on note id curating