Background:

Func `ekg-document-titles` & `ekg-active-note-ids` go through loads of
ids one-by-one to check if an id is active with `ekg-active-id-p`,
which is really expensive and takes a long time to finish for a large
database.

That makes it really struggling to use funcs that rely on the check,
for example, `ekg--transclude-titled-note-completion`, which depends
on `ekg-document-title` to refresh candidate list on every input.

--

Comparison of two functions filtering active note-ids:

Old func:

```emacs lisp
(defun old-collect-active-ids ()
  (seq-filter
   #'ekg-active-id-p
   (triples-subjects-of-type ekg-db 'titled)))
```

New func:
```emacs lisp
(defun ekg-inactive-note-ids ()
  "Get a list of ekg-note objects, representing all inactive notes.
  Inactive in this context means trashed or draft note."
  (delete-dups
   (flatten-list
    (mapcar
     (lambda (tag)
       (plist-get (triples-get-type ekg-db tag 'tag) :tagged))
     (seq-filter
      (lambda (tag)
        (string-match-p
         (rx (seq bol (or "trash/" (seq "draft" eol)))) tag))
      (triples-subjects-of-type ekg-db 'tag))))))

(defun new-collect-active-ids ()
  (seq-difference
   (triples-subjects-of-type ekg-db 'titled)
   (ekg-inactive-note-ids)))
```

--
Test: if elements filtered by the above two are the same:
```emacs lisp
(let ((old (old-collect-active-ids))
      (new (new-collect-active-ids)))
  (and (cl-subsetp old new)
       (cl-subsetp new old))) ; result => t => the elements are the same
```

--
Benchmark (rough) on my personal triples.db

record number: 32702
title number: 2401
tag number: 2566
note containing tags: 1728
note containing text: 2566

```
| fn                   | old    | new    |  ratio |
|----------------------+--------+--------+--------|
| *-collect-active-ids | 15.13s | 0.038s | 398.31 |
| ekg-document-titles  | 15.89s | 1.50s  |  10.60 |
| ekg-active-note-ids  | 15.49s | 0.039s | 397.18 |
```

As indicated above, the new method is way way way faster than the
original. So this commit re-adapts the related code to the new method.

Note: The benchmark could be biased, since my db has no more than
10 draft/trashed notes.

Update: formating