@qingshuizheng 2024-01-05
An even faster solution for `ekg--transclude-titled-note-completion` is to cache `ekg-document-titles`. But that needs some sweats, and I don't use it for the moment, 1.50 seconds (in my old macbook pro) to finish an `ekg-document-titles` call is decent enough for me after this change.