@qingshuizheng 2024-01-06
Just realize we could do the same for `ekg-show-notes-in-trash`. 
Please wait a sec, I will push another commit later.