@qingshuizheng 2024-01-06
I force-pushed a commit including:
1. `delete-dups` -> `seq-uniq`
2. change how `ekg-show-notes-in-trash` curates note ids of trashed note.