@ahyatt 2024-01-06
Thank you for these fixes, they make a lot of sense.  I think you must have the largest repository, so you seem to be running up against these perf problems first.  Let me know what else you find!